src/requirements.txt
absl-py==2.3.1
addict==2.4.0
albucore==0.0.24
albumentations==2.0.8
annotated-types==0.7.0
astunparse==1.6.3
attrs==25.3.0
basicsr==1.4.2
beautifulsoup4==4.13.4
build==1.2.2.post1
certifi==2025.7.9
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
cmake==4.0.3
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.2
cycler==0.12.1
Cython==3.1.2
dlib @ file:///D:/dlib/dist/dlib-20.0.99-cp311-cp311-win_amd64.whl#sha256=7e85017234d8acbcea526905571a2ee50987a7a95b41e0c50eabdf99e9f6bc5e
dnnlib @ https://github.com/podgorskiy/dnnlib/releases/download/0.0.1/dnnlib-0.0.1-py3-none-any.whl#sha256=e789afb6365198754f3eedd7687d2c321f6a2daf8009eb94e46fb8b90217361d
dnnlib-util==0.3.1
easydict==1.13
facexlib==0.3.0
filelock==3.18.0
filterpy==1.4.5
flatbuffers==25.2.10
fonttools==4.58.5
fsspec==2025.5.1
ftfy==6.3.1
future==1.0.0
gast==0.6.0
gdown==5.2.0
gfpgan==1.3.8
gitdb==4.0.12
GitPython==3.1.44
google-pasta==0.2.0
grpcio==1.73.1
h5py==3.14.0
humanfriendly==10.0
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.4.3
insightface==0.7.3
jax==0.6.2
jaxlib==0.6.2
Jinja2==3.1.6
joblib==1.5.1
keras==3.10.0
kiwisolver==1.4.8
lazy_loader==0.4
libclang==18.1.1
llvmlite==0.44.0
lmdb==1.6.2
lpips==0.1.4
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
mediapipe==0.10.21
ml_dtypes==0.5.1
mpmath==1.3.0
namex==0.1.0
networkx==3.5
ninja==1.11.1.4
noise==1.2.2
numba==0.61.2
numpy==2.1.3
onnx==1.18.0
onnxruntime==1.22.1
opencv-contrib-python==4.11.0.86
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
opt_einsum==3.4.0
optree==0.16.0
packaging==25.0
pillow==11.3.0
platformdirs==4.3.8
prettytable==3.16.0
protobuf==4.25.8
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pyparsing==3.2.3
pyproject_hooks==1.2.0
pyreadline3==3.5.4
PySocks==1.7.1
python-dateutil==2.9.0.post0
PyYAML==6.0.2
-e git+https://github.com/xinntao/Real-ESRGAN.git@a4abfb2979a7bbff3f69f58f58ae324608821e27#egg=realesrgan
regex==2024.11.6
requests==2.32.4
rich==14.0.0
scikit-image==0.25.2
scikit-learn==1.7.0
scipy==1.16.0
sentencepiece==0.2.0
sentry-sdk==2.32.0
simsimd==6.5.0
six==1.17.0
smmap==5.0.2
sounddevice==0.5.2
soupsieve==2.7
stringzilla==3.12.5
stylegan2-torch==1.3.0
sudo==1.0.0
sympy==1.14.0
tb-nightly==2.20.0a20250709
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.1.0
threadpoolctl==3.6.0
tifffile==2025.6.11
torch==2.7.0+cu128
torch-conv-gradfix==0.1.0
torchvision==0.22.0+cu128
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
wandb==0.21.0
wcwidth==0.2.13
Werkzeug==3.1.3
wrapt==1.17.2
yapf==0.43.0
pyspng-0.1.3

